fix: removed artifacts from gaussian generators
While testing the distribution visualizer, I noticed that the gaussian
generators were producing artifacts, 6 obvious spokes in the 2d plot.
However, these artifacts were only obvious when the underlying prng
allowed variation via a hashState function. This lead me to inspect the
tuple generator used by the gaussian generator. There I noticed the
culprit, the sequence function (and therefore the hashState function)
being called on each term of the tuple. Clearly, hashing the state of
the prng between uniforms was breaking the independence that the
gaussian generator relies on. I removed the automatic sequencing of
tuple generator terms, as well as the sequencing of records. Now,
sequence must be called explicitly. Likely a rework of the variation
system is required - it's still far from being useful in most cases.