fix: partial generator records failing to generate
Some generators were failing because they relied on the previous
behavior of the Generator.record() function when it still used
Generator.sequence to process the entries of the passed record. The
solution was to adjust the functionality of Generator.run() to more
closely match the behavior of Generator.sequence(), meaning that it will
now traverse the tree of the passed structure recursively, turning a
complex JavaScript object with potentially nested arrays and objects,
as well as generators and non-generators, into a master generator of the
entire reified structure. Since Generator.chain also uses Generator.run,
it now functions differently - not only does it act like a selective
flatMap, but it also traverses the potentially complex nested generator
structure returned by the passed chainFn, and also returns the reified
structure.

This is all quite experimental, and I'm not sure if it's great to
introduce this kind of complexity into the behavior of an innocuously-
-named function like Generator.run(), but it does seem to be quite
useful as a utility method.